+++ libs/core/datetime-picker/datetime-picker.component.ts
@@ -146,6 +146,10 @@ export class DatetimePickerComponent<D>
     @Input()
     required = false;
 
+    /** Define a custom datetime format. Can be a string for DayjsDatetimeAdapter or object for FdDatetimeAdapter  */
+    @Input()
+    customDateTimeFormat: unknown;
+
     /**
      * Whether the time component should be meridian (am/pm).
      * Default value is based on the current locale format option
@@ -819,10 +823,9 @@ export class DatetimePickerComponent<D>
      * Format date time entity.
      */
     private _formatDateTime(dateTime: D): string {
-        const formattedDate: string = this._dateTimeAdapter.format(
-            dateTime,
-            this._dateTimeFormats.display.dateTimeInput
-        );
+        const formattedDate: string = this.customDateTimeFormat
+            ? this._dateTimeAdapter.format(dateTime, this.customDateTimeFormat)
+            : this._dateTimeAdapter.format(dateTime, this._dateTimeFormats.display.dateTimeInput);
         return formattedDate || '';
     }