+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.51.3-rc.17](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.16...v0.51.3-rc.17) (2024-08-20)
+
+
+### Bug Fixes
+
+* **core:** multi input tokens should not have title attr ([#12233](https://github.com/SAP/fundamental-ngx/issues/12233)) ([8c177b0](https://github.com/SAP/fundamental-ngx/commit/8c177b0ad47a853a832c22eba5bfc727a1a2c9ea))
+
+
+
+
+
+## [0.51.3-rc.16](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.15...v0.51.3-rc.16) (2024-08-20)
+
+
+### Features
+
+* **core:** add input to customize the datetime format of specific datetime-picker component ([#12273](https://github.com/SAP/fundamental-ngx/issues/12273)) ([f45f5c3](https://github.com/SAP/fundamental-ngx/commit/f45f5c38dbe0a674c164faee66726bd42e0d36fd)), closes [#12251](https://github.com/SAP/fundamental-ngx/issues/12251)
+
+
+
+
+
 ## [0.51.3-rc.15](https://github.com/SAP/fundamental-ngx/compare/v0.51.3-rc.14...v0.51.3-rc.15) (2024-08-16)